Move NumPy specific typeof impls to NumPy module
