+++ numba/core/cpu.py
@@ -122,6 +122,7 @@ class CPUContext(BaseContext):
         from numba.typed import typedset, setobject # noqa F401
         from numba.experimental import jitclass, function_type # noqa F401
         from numba.np.types import datetime_registry # noqa F401
+        from numba.np.types import typeof # noqa F401
         from numba.np import npdatetime # noqa F401
 
         # Add target specific implementations
